requirements.txt
streamlit
streamlit-authenticator
python-dotenv
weaviate-client
qdrant-client
nomic
openai
langchain
sentence-transformers
transformers
torch
pdfplumber
tiktoken
requests
pandas
numpy
scikit-learn
faiss-cpu
uvicorn
fastapi
pydantic
beautifulsoup4
pytest
